Modify tokens, experiment with alex, add to syntax
We now lex verbatim spans slightly differently: the verbatim text and
backtick escape tokens are now single tokens at the top level, instead
of being collected into one list. The verbatim text token now does not
contain space or newline characters; those are handled with the existing
LineSpace and Indent tokens. That change should also correct the
handling of indentation in verbatim spans.

The added LexerUtils contain the definitions required for alex to lex a
document. It will serve as the foundation of an alternate lexer that may
replace the current one.

The definition of Syntax.Element has been expanded.