Start test suite, add lexing test, fix lexing bugs
The set of characters that look like they should be part of an
identifier should not include backslash. This commit properly excludes
that character from the identish set.

The number sign is insignificant if it does not occur at the beginning
of a line. The relevant character sets have been changed to allow for
this.