Add start of manual, fix indent bug and add tests
The manual is mostly a stub at this point.

The indentation stripping in level and layout elements is now based
solely on the indentation of content after the attribute set. Formerly,
the attribute set indentation was added to the content indentation and
that resulting indentation was stripped from the content.