Start refactoring scope handling
Some tests were added that are correct given our current whitespace
model. Since I now believe that that model is wrong and should be
modified slightly, these tests are going to fail and will need to be
verified to conform to the new model outlined in the commend above
doEndBlanks.