Refactor whitespace and scope handling
We reintroduce initial and terminal whitespace stripping to the various
things that create scopes.

The handling of scope closing has been greatly simplified, with a new
gatherScopesDL function removing most of the boilerplate of this task in
a far more flexible way than the replicateEndImplicitScope approach.

The lexing of blank lines has been altered slightly, according to the
principle that blank lines properly belong to the first scope that they
do not participate in closing.