Add attribute set scopes, overhaul lexing
We now track attribute set scopes, meaning that we know exactly when we
are in an attribute set now. To support this, we now keep track of
ambient start codes in the scopes so that we can restore the codes on
scope exit properly; before this change there was no real need to do
this, due to the limited number of start codes and their lack of
interactions.

Lexing of course takes advantage of this, and indeed must take advantage
of this to track attribute set scopes properly. The plain text codes
have now been merged, and new attribute-set-related codes have been
introduced. We also use this to avoid lexing '=' and insignificant
square brackets in plain text.

I added a new tokenization test for attribute sets because I had some
trouble getting the new codes and lexer working. The test should alert
us to problems in attribute set lexing early in the parsing process.

As a fun side-effect, we now only have a single shift/reduce conflict in
the grammar. I may in future eliminate this last conflict entirely.